+++ src/App.tsx
@@ -6,6 +6,7 @@ import { Trans } from "@lingui/macro";
 import { Link } from "./components/link";
 import { OfflineIndicator } from "./components/OfflineIndicator";
 import { LanguageSwitcher } from "./components/LanguageSwitcher";
+import { SyncStatusIndicator } from "./components/SyncStatusIndicator";
 
 function Home() {
   return (
@@ -52,6 +53,8 @@ function About() {
 }
 
 function App() {
+  const apiBaseUrl = import.meta.env.VITE_API_URL || "http://localhost:8000";
+
   return (
     <BrowserRouter>
       <div className="app">
@@ -61,6 +64,7 @@ function App() {
         </Routes>
       </div>
       <OfflineIndicator />
+      <SyncStatusIndicator apiBaseUrl={apiBaseUrl} />
     </BrowserRouter>
   );
 }